docs: merkle and selector docs/tests (#1152)
* docs: added docs for merkle and selector utils, moved selector to hash, added happy path tests

* chore: added example tag to jsdoc
