+++ src/index.ts
@@ -32,7 +32,7 @@ export * as typedData from './utils/typedData';
 export * as ec from './utils/ec';
 export * as starknetId from './utils/starknetId';
 export * as provider from './utils/provider';
-export * as selector from './utils/selector';
+export * as selector from './utils/hash/selector';
 export * as events from './utils/events/index';
 export * from './utils/cairoDataTypes/uint256';
 export * from './utils/cairoDataTypes/uint512';
