+++ src/utils/hash/index.ts
@@ -2,7 +2,7 @@
  * Hashes Exports
  */
 export * as poseidon from '@noble/curves/abstract/poseidon';
-export * from '../selector'; // Preserve legacy export structure
+export * from './selector'; // Preserve legacy export structure
 
 export * from './transactionHash';
 export * from './classHash';
